+++ dptb/negf/hamiltonian_init.py
@@ -1,7 +1,7 @@
 from typing import List
 import torch
 from dptb.negf.areshkin_pole_sum import pole_maker
-from dptb.negf.RGF import recursive_gf
+from dptb.negf.recursive_green_cal import recursive_gf
 from dptb.negf.surface_green import selfEnergy
 from dptb.negf.negf_utils import quad, gauss_xw,update_kmap,leggauss
 from dptb.negf.ozaki_res_cal import Ozaki_residues
